rename colors -> fmt
